Handle maxUncompactSize error case in Uncompact
The C maxUncompactSize function used in Uncompact can return
-1 as the maximum size when an error (such as the resolution being
specified is smaller than that of the hexagons being uncompacted).
This changes the function signature to return an error from
Uncompact in this case instead of the current behavior of causing
a panic when an attempt is made to make a slice of -1 size.