add support for GOMODULES
This repo follows a modified version of semver, where the major version
of this repository tracks the major version of the C core repository.
This doesn't play well with GOMODULES by default, so we create our first
major version go module that is just a duplication of master version.

Solves #24